# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple API to compile LaTeX templates and HTML documents to PDF",
        "title": "PDF Compiler",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/compile": {
            "post": {
                "description": "Compiles a LaTeX template and returns a PDF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Compile"
                ],
                "summary": "Compile LaTeX template",
                "parameters": [
                    {
                        "description": "LaTeX template",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CompileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Compilation error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/compile-html": {
            "post": {
                "description": "Compiles an HTML document and returns a PDF using headless Chrome",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Compile"
                ],
                "summary": "Compile HTML to PDF",
                "parameters": [
                    {
                        "description": "HTML document",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CompileHTMLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Compilation error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns 200 OK",
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CompileHTMLRequest": {
            "type": "object",
            "properties": {
                "html": {
                    "type": "string",
                    "example": "\u003chtml\u003e\u003cbody\u003e\u003ch1\u003eHello, world!\u003c/h1\u003e\u003c/body\u003e\u003c/html\u003e"
                }
            }
        },
        "main.CompileRequest": {
            "type": "object",
            "properties": {
                "tex": {
                    "type": "string",
                    "example": "\\documentclass{article}\n\\begin{document}\nHello, world!\n\\end{document}"
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request, must provide LaTeX template"
                }
            }
        }
    }
}